requirements.txt
websockets>=12.0
httpx>=0.26.0
anthropic>=0.18.0
python-dotenv>=1.0.0
